Lectures/assembly-examples/requirements.txt
capstone==4.0.1
pkg-resources==0.0.0
pyelftools==0.25
unicorn==1.0.1
